patch: added tokens for delete and patch users, generated dates from server side

+++ api/src/features/users/users.controller.ts
@@ -1,8 +1,14 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common'
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Req } from '@nestjs/common'
 import { ApiTags, ApiBearerAuth } from '@nestjs/swagger'
 import { UsersService } from './users.service'
 import { CreateUserDto } from './dto/create-user.dto'
 import { UpdateUserDto } from './dto/update-user.dto'
+import { AccessTokenGuard } from '../auth/guards'
+import { Request } from 'express'
+
+interface RequestWithUser extends Request {
+  user: any
+}
 
 @ApiTags('users')
 @ApiBearerAuth()
@@ -10,28 +16,40 @@ import { UpdateUserDto } from './dto/update-user.dto'
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
+  @UseGuards(AccessTokenGuard)
   @Post()
   create(@Body() createUserDto: CreateUserDto) {
     return this.usersService.create(createUserDto)
   }
 
+  @UseGuards(AccessTokenGuard)
   @Get()
   findAll() {
     return this.usersService.findAll()
   }
 
+  @UseGuards(AccessTokenGuard)
   @Get(':id')
   findOne(@Param('id') id: string) {
     return this.usersService.findOne(id)
   }
 
+  @UseGuards(AccessTokenGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto, @Req() req: RequestWithUser) {
+    if(req.user.sub !== id) {
+      throw new Error('You are not authorized to update this user')
+    }
+
     return this.usersService.update(id, updateUserDto)
   }
 
+  @UseGuards(AccessTokenGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @Req() req: RequestWithUser) {
+    if(req.user.sub !== id) {
+      throw new Error('You are not authorized to delete this user')
+    }
     return this.usersService.remove(id)
   }
 }
